requirements.txt
api.ai==0.1.0
apiai==1.2.3
bidict==0.21.4
cachelib==0.4.1
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.9
click==8.0.3
Flask==2.0.2
Flask-Session==0.4.0
Flask-SocketIO==5.1.1
gevent==21.12.0
gevent-socketio==0.3.6
gevent-websocket==0.10.1
google-api-core==2.3.2
google-auth==2.3.3
google-cloud-dialogflow==2.10.0
googleapis-common-protos==1.54.0
greenlet==1.1.2
grpcio==1.43.0
grpcio-status==1.43.0
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
monotonic==1.6
numpy==1.21.4
proto-plus==1.19.8
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
python-engineio==4.3.0
python-socketio==5.5.0
requests==2.26.0
rsa==4.8
six==1.16.0
urllib3==1.26.7
Werkzeug==2.0.2
zope.event==4.5.0
zope.interface==5.4.0
